Add corrected color to filtered scores and move to utils
